significant changes across the board
first of all, i renamed `gefjon-utils:defclass` to `define-class` to reduce
heinousness.

second, i refactored [`cps/expr`] so that all binding forms (not let, which is
actually more of an operator-applying form) are the expr `bind`, which holds a
single subclass of `definition`. this introduced a relatively complex class
hierarchy for `definition`, since `cont` and `func` are more similar to each
other than they are to `const`. this increases complexity in the
representation, but decreases complexity in the transformation code. also, i
hope that it will ease future refactoring to support global bindings.

i further refactored [`cps/trans`] s.t. `access-closure-env` is now a kind of
`definition`, rather than an `operator`. this allows it to take its index
directly as an argument, rather than requiring it be bound in a `const` first.