fix bugs
most notably, this commit fixes a bug where closure variables from
inner scopes weren't being propagated to intermediate scopes, leading
to an attempt to call `make-closure-env` on variables not it scope.

it also fixes a small bug in the cps->3adr transform for primitive
operators which had failed to transform subsequent forms.